@@ -5,23 +5,19 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.4.RELEASE</version>
+        <version>2.4.1</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.github.sankowskiwojciech</groupId>
     <artifactId>courses-lessons</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>courses-lessons</name>
-    <description>Demo project for Spring Boot</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>15</java.version>
 
         <courses-core-lib.version>1.0-SNAPSHOT</courses-core-lib.version>
 
-        <commons-collections4.version>4.4</commons-collections4.version>
-        <commons-lang3.version>3.10</commons-lang3.version>
-        <lombok.version>1.18.16</lombok.version>
         <tika-core.version>1.24.1</tika-core.version>
         <tika-parsers.version>1.24.1</tika-parsers.version>
     </properties>
@@ -35,7 +31,6 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>${lombok.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -74,16 +69,6 @@
             <groupId>org.apache.tika</groupId>
             <artifactId>tika-parsers</artifactId>
             <version>${tika-parsers.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>${commons-lang3.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-collections4</artifactId>
-            <version>${commons-collections4.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
